[central] Update workspace if directory has changed
The existing code assumed that the ws would always be changing direct from
default to a cnf workspace, or direct from a cnf workspace to a default.
This is most probably the case nearly all the time. However, just to be a
bit more defensive, put code in here to handle the case where we change
direct from one cnf to another.

Signed-off-by: Fr Jeremy Krieg <fr.jkrieg@greekwelfaresa.org.au>